requirement.txt
astroid==2.4.2
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
opencv-python==4.4.0.42
Pillow==7.2.0
pylint==2.6.0
python-osc==1.7.4
six==1.15.0
toml==0.10.1
wrapt==1.12.1
